Move the two `*ReconstructionContext` classes into `wire-common`.
