+++ packages/data-model/src/wire-common/BaseReconstructionContext.ts
@@ -4,8 +4,8 @@
  * via a single shared implementation instead of repeating it.
  */
 
-import type { FabricInstance } from "./interface.ts";
-import type { ReconstructionContext } from "./wire-common/interface.ts";
+import type { FabricInstance } from "../interface.ts";
+import type { ReconstructionContext } from "./interface.ts";
 
 /**
  * Abstract base that supplies the `shouldDeepFreeze` getter from a
