+++ packages/data-model/src/wire-common/index.ts
@@ -10,3 +10,8 @@ export {
 
 export { WIRE_META_TAGS } from "./wire-meta-tags.ts";
 export { WIRE_TYPE_TAGS } from "./wire-type-tags.ts";
+export { BaseReconstructionContext } from "./BaseReconstructionContext.ts";
+export {
+  EMPTY_RECONSTRUCTION_CONTEXT,
+  EmptyReconstructionContext,
+} from "./EmptyReconstructionContext.ts";
