+++ packages/data-model/test/json-wire/json-encoding.test.ts
@@ -8,7 +8,7 @@ import {
 } from "../../src/json-wire/json-encoding.ts";
 import { FabricError } from "../../src/fabric-instances/FabricError.ts";
 import type { FabricValue } from "../../src/fabric-value.ts";
-import { BaseReconstructionContext } from "../../src/BaseReconstructionContext.ts";
+import { BaseReconstructionContext } from "../../src/wire-common/BaseReconstructionContext.ts";
 
 /** Mock runtime for deserialization calls. */
 class MockRuntime extends BaseReconstructionContext {
